Added basic structure for peripheral
:small_blue_diamond: Added peripheral block complete with tile entity
and that fun stuff (first time modding :smile:).
:small_blue_diamond: Added ChatterBotAPI (credits to Pierre-David
Bélanger on github for awesome library) to the project so it's possible
to play around with Cleverbot :grin:.
:small_blue_diamond: Added peripheral methods that hooks up quite neatly
to the ChatterBotAPI.

:+1: Played around with the peripheral and it works pretty good so far.
Maybe a tad of latency every now and then but other than that it does
what it should do.

:small_red_triangle: TODO: change out the UUID for sessions to instead
have session objects. For a more pleasing user experience.
:small_red_triangle: TODO: create a texture for it :sweat_smile: