Final changes to the source (i think)...
:small_blue_diamond: Moved the AIChatRequest.java file into the
cleverbot package.. (idk where to put the file :joy:)
:small_blue_diamond: Using jar file instead of maven, so I can bake it
into the final P++ jar. (Instead of having to have it externally)
:small_blue_diamond: Added methods to get a session for a certain box.
This way you can have multiple computers hooked up to the same session
(if you for some reason wanted that... :smirk:).

:small_orange_diamond: Hopefully the P++ guys appriciate my work
:cold_sweat:.

:small_red_triangle: TODO: Writing docs for peripheral.

(NOTE TO SELF: DONT FORGET TO GIVE CREDITS TO THE CREATOR OF THE
CHATTER-BOT-API)